// js/types.ts
export interface SCUser {
  id?: number;
  username: string;
  avatar_url?: string | null;
}
// ...
export interface TokenPayload {
  accessToken?: string;
  access_token?: string;
  refreshToken?: string;
  refresh_token?: string;
  expiresIn?: number | string;
  expires_in?: number | string;
}